Update index, script, styles
Add a pages section to the grid. In the script, add code to stop users from submitting form without any inputs. Add a read or not read radio buttons and link them to the grid in the html. Update styles for all the new elements added and update existing elements.